Add files via upload
Developed a classification model in python to predict if a patient has Diabetes or not based on medical features such as glucose intake, insulin level, blood pressure, body mass index (BMI) with an accuracy of 83% and recall of 68% on the selected random forest classifier model. Libraries used: Pandas, Numpy, Matplotlib, Seaborn, Scikit-learn.